JWT-JWK/requirements.txt
jwcrypto
pyjwt
requests